noexcept

+++ bits_test.cc
@@ -1,3 +1,4 @@
+#include <cstdint>
 #include <gtest/gtest.h>
 #include "bits.h"
 
@@ -5,7 +6,7 @@ namespace {
 TEST(bits, merge) {
   auto a = 0b00000000000000000000000000000000U;
   auto b = 0b11111111111111111111111111111111U;
-  auto mask = fmath::bitmask<unsigned int>(16);
+  auto mask = fmath::bitmask<uint64_t>(16);
   ASSERT_EQ(fmath::merge(a, b, mask), mask);
 
 }
